requirements.txt
pymupdf
pytesseract
pdf2image
sqlalchemy
psycopg2-binary
playwright
numpy
tabula_py
pandas